requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-google-genai>=1.0.0
langchain-community>=0.0.20
faiss-cpu>=1.7.4
pypdf>=3.17.0
unstructured>=0.11.0
python-dotenv>=1.0.0
openai>=1.12.0
google-generativeai>=0.3.0
tiktoken>=0.5.0
numpy>=1.24.0
pandas>=2.0.0